Rename main package to benchmark-runner and update readme
