+++ README.md
@@ -63,3 +63,14 @@ _Requires python 3.9._
 - Install dependencies: `pip install -r requirements.txt`
 - Start the frontend server: `python server.py`
 - Open in your browser
+
+## Benchmark Runner
+The benchmarks can be installed from pypi and used as part of the Translator-wide automated testing.
+- `pip install benchmarks-runner`
+To run benchmarks:
+```python
+from benchmarks_runner import run_benchmarks
+
+run_benchmarks(<benchmark>, <target>)
+```
+where benchmark is the name of a benchmark that is specified in config/benchmarks.json, and a target that is specified in config/targets.json
