+++ benchmarks_runner/cli/eval.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import os
 
 
-from benchmarks.eval import evaluate_results
+from benchmarks_runner.eval import evaluate_results
 
 metrics_at_k = {
     'Precision @ k\t\t': 'precision_at_k',
