+++ benchmarks_runner/eval.py
@@ -11,7 +11,7 @@ import numpy as np
 from matplotlib import pyplot as plt
 from matplotlib.axes import Axes
 
-from benchmarks.utils.benchmark import benchmark_ground_truth
+from benchmarks_runner.utils.benchmark import benchmark_ground_truth
 
 
 def evaluate_results(
